requirements.txt
absl-py==1.0.0
asgiref==3.5.0
astunparse==1.6.3
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cloudinary==1.29.0
Django==4.0.3
django-environ==0.8.1
djangorestframework==3.13.1
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.45.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
pathspec==0.9.0
platformdirs==2.4.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io==0.24.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.9
Werkzeug==2.0.3
whitenoise==6.0.0
wrapt==1.14.0
zipp==3.7.0
